refactor: BaseContentEntity 삭제
BaseContentEntity의 필요성에 대한 고찰을 해본 결과, 각각의 속성에 따라 역할이 다르게 확장될 가능성이 있는 것들은 각자의 Entity로 다시 복원시켰고, Content는 분리하는 의미가 다소 부족해서 결과적으론 BaseContentEntity를 삭제하는 것으로 변경하였습니다.
